Deep-learning/Wildprint Classifier/app/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.9.0
click==8.2.1
fastapi==0.116.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h5py==3.14.0
iniconfig==2.1.0
itsdangerous==2.2.0
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
mdurl==0.1.2
mkl-service==2.4.0
ml_dtypes==0.5.3
namex==0.1.0
opt_einsum==3.4.0
optree==0.17.0
pillow==11.3.0
pluggy==1.6.0
protobuf==6.32.0
pydantic==2.11.9
pydantic_core==2.33.2
PyQt6==6.7.1
pytest==8.4.2
python-multipart==0.0.20
rich==14.1.0
setuptools==72.1.0
SQLAlchemy==2.0.43
starlette==0.47.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
typing-inspection==0.4.1
uvicorn==0.35.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
